adjust the mod use statement
