+++ core/src/sync/state/mod.rs
@@ -2,6 +2,7 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
+mod restore;
 mod snapshot_chunk_request;
 mod snapshot_chunk_response;
 mod snapshot_chunk_sync;
@@ -9,7 +10,6 @@ mod snapshot_manifest_request;
 mod snapshot_manifest_response;
 mod storage;
 
-use self::storage::{Chunk, ChunkKey, RestoreProgress, Restorer};
 pub use self::{
     snapshot_chunk_request::SnapshotChunkRequest,
     snapshot_chunk_response::SnapshotChunkResponse,
